business-backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pymysql
cryptography
httpx
python-jose
passlib
python-multipart
python-dotenv
pytest 
cos-python-sdk-v5
aiohttp
python-multipart
pydantic-settings
PyJWT
httpx
Pillow